bazucando o steam

+++ entrypoint.sh
@@ -10,8 +10,12 @@ pip install hatch
 hatch build
 pip install -e .
 apt install ./google-chrome.deb -y
-ln -s /tiktok/chrome/chromedriver/chromedriver /usr/local/bin/chromedriver
-chmod +x /usr/local/bin/chromedriver~
+
+if [ ! -L "/usr/local/bin/chromedriver" ]; then
+    ln -s /tiktok/chrome/chromedriver/chromedriver /usr/local/bin/chromedriver
+fi
+
+chmod +x /usr/local/bin/chromedriver
 
 # install coisas steam
 cd /steam
